fixed some warning
and added more warning by default

+++ include/autotune/continuous_parameter.hpp
@@ -154,7 +154,7 @@ public:
 
   size_t count_values() const {
     // TODO: implement
-    return std::floor((max - min) / step) + 1;
+    return static_cast<size_t>(std::floor((max - min) / step)) + 1;
   }
 };
