+++ include/autotune/tuners/monte_carlo.hpp
@@ -52,16 +52,17 @@ public:
         auto &p = current_parameters[parameter_index];
 
         if (p->is_integer_parameter()) {
-	  std::uniform_int_distribution<size_t> distribution(static_cast<size_t>(p->get_min()),
-							     static_cast<size_t>(p->get_max()));
-	  std::random_device rd;
-	  std::default_random_engine generator(rd());
-          p->set_value(distribution(generator));
+          std::uniform_int_distribution<size_t> distribution(
+              static_cast<size_t>(p->get_min()),
+              static_cast<size_t>(p->get_max()));
+          std::random_device rd;
+          std::default_random_engine generator(rd());
+          p->set_value(static_cast<double>(distribution(generator)));
         } else {
-	  std::uniform_real_distribution<double> distribution(p->get_min(),
-					     		      p->get_max());
-	  std::random_device rd;
-	  std::default_random_engine generator(rd());
+          std::uniform_real_distribution<double> distribution(p->get_min(),
+                                                              p->get_max());
+          std::random_device rd;
+          std::default_random_engine generator(rd());
           p->set_value(distribution(generator));
         }
       }
